Implemented a blackbox check
