+++ eclipse-workspace/halstead/src/main/java/loopingStatementsPackage/LoopingStatementsCheck.java
@@ -1,5 +1,7 @@
 package loopingStatementsPackage;
 
+import java.util.Hashtable;
+
 import com.puppycrawl.tools.checkstyle.api.AbstractCheck;
 import com.puppycrawl.tools.checkstyle.api.DetailAST;
 
@@ -51,5 +53,12 @@ public class LoopingStatementsCheck extends AbstractCheck {
 	{
 		log(aAST.getLineNo(), message + count + " -AJ");
 		count = 0;
+	}
+
+	public Hashtable<String, Integer> getResults() 
+	{
+		Hashtable<String, Integer> result = new Hashtable<String, Integer>();
+		result.put(this.message, this.count);
+		return result;
 	}
 }
